# docs/swagger.yaml
components:
  schemas:
    apperrors.AppError:
      properties:
        code:
          type: string
        message:
          type: string
        params:
          additionalProperties: {}
          type: object
      type: object
...
    database.VersionAuthor:
      properties:
        kind:
          description: 'operator: written in SkyMail; template_seed: written by the
            Template seed from the repo.'
          enum:
          - operator
          - template_seed
          type: string
        name:
          description: The name the writer's token carried when the version was written.
            Null when not known.
          type: string
        sub:
          description: 'The Keycloak subject of the token that wrote the version.
            Null when not known: the first versions, made from templates written before
            versions were kept.'
          type: string
      type: object
...
    handlers.Template:
      properties:
        archived_at:
          type: string
        archived_by:
          type: string
        created_at:
          type: string
        drafts:
          description: 'Each operator''s draft in progress, newest first: the newest
            version an operator wrote of the template, while it is unpublished. An
            operator''s earlier drafts are superseded and not listed. A draft whose
            base_version_id is not published_version_id is stale: a newer version
            was published after it was started, and publishing it takes force.'
          items:
            $ref: '#/components/schemas/handlers.TemplateVersionSummary'
          type: array
          uniqueItems: false
        html_content:
          type: string
        id:
          type: string
        key:
          type: string
        main_mode:
          description: 'The Authoring mode of the Main source the template sends:
            its published version''s. Null for a template with no published version.'
          enum:
          - jsx
          - visual
          - html
          type: string
        name:
          type: string
        plain_text_content:
          type: string
        published_version_id:
          type: string
        react_email_content:
          type: string
        subject:
          type: string
        system:
          type: boolean
        updated_at:
          type: string
      type: object
...
    handlers.TemplateVersionSummary:
      properties:
        author:
          $ref: '#/components/schemas/database.VersionAuthor'
        base_version_id:
          description: The published version a draft started from; null for a template's
            first version.
          type: string
        created_at:
          type: string
        current:
          description: Whether this is the published version the template sends.
          type: boolean
        id:
          type: string
        main_mode:
          description: 'The Authoring mode whose source is the Main source: its render
            is what the version sends.'
          enum:
          - jsx
          - visual
          - html
          type: string
        published_at:
          description: When the version was published; null for a draft.
          type: string
        requested_subject:
          description: 'The subject the Template seed sent for its version. It can
            differ from subject: until the seed''s conflict rule, the seed keeps the
            subject a template already has. Null on an operator''s version and on
            the first versions, made from templates written before versions were kept.'
          type: string
        seq:
          description: 1, 2, 3… within the template, in the order versions were written.
          type: integer
        subject:
          type: string
        template_id:
          type: string
      type: object
...
    requests.CreateTemplate:
      properties:
        html_content:
          type: string
        key:
          type: string
        name:
          type: string
        plain_text_content:
          type: string
        react_email_content:
          type: string
        subject:
          type: string
      required:
      - html_content
      - name
      - plain_text_content
      - react_email_content
      - subject
      type: object
...
    requests.UpdateTemplate:
      properties:
        html_content:
          type: string
        key:
          type: string
        name:
          type: string
        plain_text_content:
          type: string
        react_email_content:
          type: string
        subject:
          type: string
      required:
      - html_content
      - name
      - plain_text_content
      - react_email_content
      - subject
      type: object
    requests.UpsertTemplateByKey:
      properties:
        html_content:
          type: string
        name:
          type: string
        plain_text_content:
          type: string
        react_email_content:
          type: string
        subject:
          type: string
        system:
          type: boolean
      required:
      - html_content
      - name
      - plain_text_content
      - react_email_content
      - subject
      type: object
...
info:
  contact:
    email: hello@enesgenc.dev
    name: Enes Genç
    url: https://enesgenc.dev
  description: This is the API documentation for Skymail.
  title: Skymail
  version: "1.0"
openapi: 3.1.0
paths:
...
  /templates:
    get:
      description: Get a list of all email templates with pagination.
      parameters:
      - description: Start index
        in: query
        name: _start
        schema:
          type: integer
      - description: End index
        in: query
        name: _end
        schema:
          type: integer
      - description: 'Lifecycle filter: current, inactive, all'
        in: query
        name: lifecycle
        schema:
          default: current
          enum:
          - current
          - inactive
          - all
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/handlers.Template'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: List all email templates
      tags:
      - Templates
    post:
      description: 'Create a new email template with the provided name, HTML content,
        and plain text content. Records the content as the template''s first version:
        an operator''s, published at once.'
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.CreateTemplate'
                description: Template details
                summary: template
        description: Template details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Create a new email template
      tags:
      - Templates
  /templates/{id}:
    delete:
      description: Archive an email template without removing historical mail tasks
        or queue items. Repeating the request is safe.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Archive an email template
      tags:
      - Templates
    get:
      description: Get details of a specific email template by its ID.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Get an email template by ID
      tags:
      - Templates
    patch:
      description: Update an existing email template with the provided ID and details.
        Records the content the template ends up with as an operator's version, published
        at once.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.UpdateTemplate'
                description: Template details
                summary: template
        description: Template details
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Update an email template
      tags:
      - Templates
...
  /templates/{id}/restore:
    post:
      description: Restore an archived email template. Repeating the request is safe.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Conflict
      summary: Restore an archived email template
      tags:
      - Templates
...
  /templates/by-key/{key}:
    get:
      description: Get a template by its stable key (for example keycloak.verify-email).
      parameters:
      - description: Template key
        in: path
        name: key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Not Found
      summary: Get an email template by key
      tags:
      - Templates
    put:
      description: 'Seed path for system templates: creates the template when the
        key is new and replaces its content when it already exists. Un-archives the
        template so a seed always leaves a usable template behind. Records the content
        the template ends up with as a Template seed version, published at once.'
      parameters:
      - description: Template key
        in: path
        name: key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/requests.UpsertTemplateByKey'
                description: Template details
                summary: template
        description: Template details
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.Template'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/apperrors.AppError'
          description: Internal Server Error
      summary: Create or replace a template addressed by key
      tags:
      - Templates
servers:
- url: skymail-api.yildizskylab.com/v1
...